# docs/apidog.openapi.compat.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MLE Home Backend API",
    "version": "0.1.0",
    "description": "OpenAPI import file for Apidog. This describes the Dockerized local API for student linking, diary records, evidence uploads, OCR inference, privacy decisions, tutor summaries, reports, curriculum standards, and integration metadata."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Docker Compose API"
    }
  ],
...
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login with email and password",
        "operationId": "loginWithPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "parent@example.com",
                "password": "correct horse password"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token set.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenSetResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh and rotate tokens",
        "operationId": "refreshAuthTokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              },
              "example": {
                "refreshToken": "paste-refresh-token-here"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New token set.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenSetResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout one session by revoking a refresh token",
        "operationId": "logout",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              },
              "example": {
                "refreshToken": "paste-refresh-token-here"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refresh token revoked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/logout-all": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout all sessions for the authenticated account",
        "operationId": "logoutAll",
        "parameters": [
          {
            "$ref": "#/components/parameters/AuthorizationHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "Refresh tokens revoked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutAllResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/auth/student-links": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Create a student link code",
        "operationId": "createStudentLink",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountIdHeader"
          },
          {
            "$ref": "#/components/parameters/AccountRoleHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentLinkRequest"
              },
              "example": {
                "studentId": "student_demo",
                "scopes": [
                  "stats:read",
                  "entries:create"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Student link code created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentLinkResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      },
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "List student links for current account",
        "operationId": "listStudentLinks",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountIdHeader"
          },
          {
            "$ref": "#/components/parameters/AccountRoleHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "Student links.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentLinkListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/student-links/activate": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Activate a student link code",
        "operationId": "activateStudentLink",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountIdHeader"
          },
          {
            "$ref": "#/components/parameters/AccountRoleHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateStudentLinkRequest"
              },
              "example": {
                "code": "07B2FD7A"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student link activated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentLinkResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/v1/diary/entries": {
      "post": {
        "tags": [
          "Diary"
        ],
        "summary": "Create a diary entry",
        "operationId": "createDiaryEntry",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountIdHeader"
          },
          {
            "$ref": "#/components/parameters/AccountRoleHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDiaryEntryRequest"
              },
              "example": {
                "studentId": "student_demo",
                "durationMinutes": 60,
                "subject": "maths",
                "title": "Fractions practice",
                "description": "Completed worksheet on equivalent fractions.",
                "curriculumTags": [
                  {
                    "subject": "maths",
                    "keyStage": "KS2",
                    "topic": "Fractions",
                    "confidence": 0.8
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Diary entry created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiaryEntryResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/v1/diary/students/{studentId}/entries": {
      "get": {
        "tags": [
          "Diary"
        ],
        "summary": "List diary entries for a student",
        "operationId": "listDiaryEntries",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountIdHeader"
          },
          {
            "$ref": "#/components/parameters/AccountRoleHeader"
          },
          {
            "$ref": "#/components/parameters/StudentIdPath"
          },
          {
            "$ref": "#/components/parameters/FromQuery"
          },
          {
            "$ref": "#/components/parameters/ToQuery"
          },
          {
            "$ref": "#/components/parameters/LimitQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "Diary entries.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiaryEntryListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/diary/students/{studentId}/stats": {
      "get": {
        "tags": [
          "Diary"
        ],
        "summary": "Get diary statistics for a student",
        "operationId": "getDiaryStats",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountIdHeader"
          },
          {
            "$ref": "#/components/parameters/AccountRoleHeader"
          },
          {
            "$ref": "#/components/parameters/StudentIdPath"
          },
          {
            "$ref": "#/components/parameters/FromQuery"
          },
          {
            "$ref": "#/components/parameters/ToQuery"
          },
          {
            "$ref": "#/components/parameters/LimitQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "Diary statistics.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiaryStatsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "AuthorizationHeader": {
        "name": "Authorization",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "example": "Bearer paste-access-token-here"
        }
      },
      "AccountIdHeader": {
        "name": "x-account-id",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string",
          "example": "parent_demo"
        }
      },
      "AccountRoleHeader": {
        "name": "x-account-role",
        "in": "header",
        "required": false,
        "schema": {
          "$ref": "#/components/schemas/AccountRole"
        }
      },
      "StudentIdPath": {
        "name": "studentId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "example": "student_demo"
        }
      },
...
      "FromQuery": {
        "name": "from",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "ToQuery": {
        "name": "to",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "LimitQuery": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 500,
          "default": 50
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication failed or is required.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "AccountRole": {
        "type": "string",
        "enum": [
          "parent",
          "child",
          "tutor",
          "integration"
        ],
        "example": "parent"
      },
      "Subject": {
        "type": "string",
        "enum": [
          "maths",
          "english",
          "science",
          "pe",
          "art",
          "humanities",
          "life-skills",
          "other"
        ]
      },
      "LinkScope": {
        "type": "string",
        "enum": [
          "stats:read",
          "entries:create",
          "evidence:create",
          "reports:tutor"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "example": "Request validation failed."
              },
              "details": {
                "type": "object"
              },
              "issues": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "parent@example.com"
          },
          "password": {
            "type": "string",
            "example": "correct horse password"
          }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "required": [
          "refreshToken"
        ],
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "paste-refresh-token-here"
          }
        }
      },
      "PublicAccount": {
        "type": "object",
        "required": [
          "_id",
          "email",
          "role",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "example": "acct_123"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "parent@example.com"
          },
          "displayName": {
            "type": "string",
            "example": "Demo Parent"
          },
          "role": {
            "$ref": "#/components/schemas/AccountRole"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "AuthTokenSetResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "account",
              "accessToken",
              "accessTokenExpiresAt",
              "refreshToken",
              "refreshTokenExpiresAt",
              "tokenType"
            ],
            "properties": {
              "account": {
                "$ref": "#/components/schemas/PublicAccount"
              },
              "accessToken": {
                "type": "string"
              },
              "accessTokenExpiresAt": {
                "type": "string",
                "format": "date-time"
              },
              "refreshToken": {
                "type": "string"
              },
              "refreshTokenExpiresAt": {
                "type": "string",
                "format": "date-time"
              },
              "tokenType": {
                "type": "string",
                "example": "Bearer"
              }
            }
          }
        }
      },
      "LogoutResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "revoked": {
                "type": "boolean",
                "example": true
              }
            }
          }
        }
      },
      "LogoutAllResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "revokedCount": {
                "type": "integer",
                "example": 3
              }
            }
          }
        }
      },
      "CreateStudentLinkRequest": {
        "type": "object",
        "required": [
          "studentId"
        ],
        "properties": {
          "studentId": {
            "type": "string",
            "example": "student_demo"
          },
          "scopes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LinkScope"
            },
            "default": [
              "stats:read",
              "entries:create"
            ]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ActivateStudentLinkRequest": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "07B2FD7A"
          }
        }
      },
      "StudentLink": {
        "type": "object",
        "required": [
          "_id",
          "studentId",
          "guardianAccountId",
          "code",
          "scopes",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "example": "link_9444e378-9e60-40e1-923a-86c939a4d10f"
          },
          "studentId": {
            "type": "string",
            "example": "student_demo"
          },
          "guardianAccountId": {
            "type": "string",
            "example": "parent_demo"
          },
          "tutorAccountId": {
            "type": "string",
            "example": "tutor_demo"
          },
          "code": {
            "type": "string",
            "example": "07B2FD7A"
          },
          "scopes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LinkScope"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "active",
              "revoked"
            ]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "activatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "StudentLinkResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/StudentLink"
          }
        }
      },
      "StudentLinkListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentLink"
            }
          }
        }
      },
      "CurriculumTag": {
        "type": "object",
        "required": [
          "subject"
        ],
        "properties": {
          "standardId": {
            "type": "string"
          },
          "keyStage": {
            "type": "string",
            "example": "KS2"
          },
          "subject": {
            "type": "string",
            "example": "maths"
          },
          "topic": {
            "type": "string",
            "example": "Fractions"
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "example": 0.8
          }
        }
      },
      "EvidenceReference": {
        "type": "object",
        "required": [
          "evidenceId",
          "includeInGuardianReport",
          "includeImageInReports"
        ],
        "properties": {
          "evidenceId": {
            "type": "string"
          },
          "objectKey": {
            "type": "string"
          },
          "mimeType": {
            "type": "string",
            "example": "image/jpeg"
          },
          "description": {
            "type": "string"
          },
          "includeInGuardianReport": {
            "type": "boolean",
            "default": true
          },
          "includeImageInReports": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "CreateDiaryEntryRequest": {
        "type": "object",
        "required": [
          "studentId",
          "durationMinutes",
          "subject",
          "title"
        ],
        "properties": {
          "studentId": {
            "type": "string",
            "example": "student_demo"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "durationMinutes": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1440,
            "example": 60
          },
          "subject": {
            "$ref": "#/components/schemas/Subject"
          },
          "title": {
            "type": "string",
            "maxLength": 160,
            "example": "Fractions practice"
          },
          "description": {
            "type": "string",
            "maxLength": 4000
          },
          "location": {
            "type": "string",
            "maxLength": 240
          },
          "tutorAccountId": {
            "type": "string"
          },
          "evidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            },
            "default": []
          },
          "curriculumTags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CurriculumTag"
            },
            "default": []
          },
          "shareWithTutorIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          }
        }
      },
...
      "DiaryEntryResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DiaryEntry"
          }
        }
      },
      "DiaryEntryListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiaryEntry"
            }
          }
        }
      },
...
      "DiaryStatsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DiaryStats"
          }
        }
      },
...
    }
  }
}